pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "enterprise-adk"
version = "0.1.0"
description = "Enterprise AI Dev Kit — Databricks AI Dev Kit + enterprise skills for Claude"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "typer[all]>=0.12.0",
    "rich>=13.7.0",
    "pydantic>=2.0",
    "pyyaml>=6.0",
    "questionary>=2.0",
]

[project.scripts]
enterprise-adk = "enterprise_adk.cli:main"

[tool.hatch.build.hooks.custom]
path = "hatch_build.py"

[tool.hatch.build.targets.wheel]
packages = ["enterprise_adk", "enterprise_skills"]

[tool.hatch.build]
include = [
    "enterprise_adk/**",
    "enterprise_skills/**",
]
